chore: fix lint warnings

+++ contracts/Tokens/Prime/Interfaces/IVToken.sol
@@ -2,19 +2,19 @@
 pragma solidity 0.8.13;
 
 interface IVToken {
-    function borrowBalanceStored(address account) external view returns (uint);
+    function borrowBalanceStored(address account) external view returns (uint256);
 
-    function exchangeRateStored() external view returns (uint);
+    function exchangeRateStored() external view returns (uint256);
 
-    function balanceOf(address account) external view returns (uint);
+    function balanceOf(address account) external view returns (uint256);
 
     function underlying() external view returns (address);
 
-    function totalBorrows() external view returns (uint);
+    function totalBorrows() external view returns (uint256);
 
-    function borrowRatePerBlock() external view returns (uint);
+    function borrowRatePerBlock() external view returns (uint256);
 
-    function reserveFactorMantissa() external view returns (uint);
+    function reserveFactorMantissa() external view returns (uint256);
 
     function decimals() external view returns (uint8);
 }
